Move plotting functions to gammapy.visualization
