+++ gammapy/datasets/spectrum.py
@@ -17,7 +17,6 @@ from gammapy.utils.scripts import make_name, make_path
 __all__ = [
     "SpectrumDatasetOnOff",
     "SpectrumDataset",
-    "plot_spectrum_datasets_off_regions",
     "SpectrumEvaluator",
 ]
 
@@ -1316,34 +1315,6 @@ def _read_ogip_hdulist(
         region=region,
         wcs=wcs,
     )
-
-
-def plot_spectrum_datasets_off_regions(datasets, ax=None):
-    """Plot spectrum datasets of regions.
-
-    Parameters
-    ----------
-    datasets : list of `SpectrumDatasetOnOff`
-        List of spectrum on-off datasets
-    """
-    import matplotlib.pyplot as plt
-    import matplotlib.patches as mpatches
-
-    ax = plt.gca() or ax
-
-    color_cycle = plt.rcParams["axes.prop_cycle"]
-    colors = color_cycle.by_key()["color"]
-    handles = []
-
-    for color, dataset in zip(colors, datasets):
-        kwargs = {"edgecolor": color, "facecolor": "none"}
-        dataset.counts_off.plot_region(ax=ax, **kwargs)
-
-        # create proxy artist for the custom legend
-        handle = mpatches.Patch(label=dataset.name, **kwargs)
-        handles.append(handle)
-
-    plt.legend(handles=handles)
 
 
 class SpectrumEvaluator:
