+++ gammapy/visualization/__init__.py
@@ -1,0 +1,3 @@
+from .cmap import *
+from .panel import *
+from .utils import *
